QSS17 Project1
For this project, I am using the data from the 35th Congress (1857-1859) and the 116th Congress (2019-2021). According to VoteView, the 35th Congress was the first time the Republican Party was part of the Congress. I want to compare the number of democrats and republicans from the 35th Congress and 116th Congress. I also want to see how the First Dimension Estimates (nominate_dim1) compare within the two parties among the different periods.